+++ NEWS.md
@@ -7,6 +7,7 @@
 * `calc_taxon_abund` no longer errors when a taxon has no observations associated with it.
 * New `heat_tree_matrix` options to change size and color of row and column labels. 
 * Fixed a bug causing the size legend not to be shown (issue [#249](https://github.com/grunwaldlab/metacoder/issues/249).
+* Now when a node_color_interval is set but a edge_color_interval is not and edge_color is not used, the edge_color_interval is the same as the node_color_interval.
 
 ### New features