+++ NEWS.md
@@ -13,6 +13,10 @@
 
 * Added parser for `dada2` results called `parse_dada2` and writers to convert back called `make_dada2_asv_table` and `make_dada2_tax_table`.
 
+### Changes
+
+* Started using the viridis colors for `heat_tree` by default (issue [#133](https://github.com/grunwaldlab/metacoder/issues/133).
+
 ##  metacoder 0.3.0
 
 ### Bug fixes