+++ NEWS.md
@@ -1,6 +1,6 @@
 # News 
 
-## Current
+##  metacoder 0.3.1
 
 ### Bug fixes